fix(disas/db) explicitly set ida database file name
ida64 6.9 (tested on linux so far) removes the original file extension from the generated database path.
Example: `/some/path/foobar.so` --> `/some/path/foobar.i64`. This causes the analysis script to be executed on a non-existing ida database (`/some/path/foobar.so.i64`).

This commit changes so that the generated database name is specified as an output database name (`-o`) in the createDatabase command.