[Bug Fix] Enforce logic applies only to active code types
Fixes the main bug in issue #41. When there was only 1 active code
type (THUMB) and we were scanning a line from the other code type (ARM)
we triggered an exception.

Scanned the code to make sure no prediction will be made on code
types which are not active at the moment of the decision.